Moved the underscore code to air.js for more modularity
